Refactoring all payTrans functions into payTransFunctions.py
